api pamdes

+++ database/migrations/2020_07_05_130549_create_pelanggans_table.php
@@ -16,12 +16,13 @@ class CreatePelanggansTable extends Migration
         Schema::create('pelanggans', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('nomor_telepon');
+            $table->bigInteger('nomor_telepon');
             $table->string('jenis_kelamin');
-            $table->integer('nomor_kk');
+            $table->bigInteger('nomor_kk');
             $table->integer('id_dusun');
             $table->text('alamat');
             $table->timestamps();
+            
         });
     }
